Rename solana contract to solana program
