+++ identity-service/.gitignore
@@ -11,7 +11,7 @@ local-storage/
 local-test/
 contract-config.json
 eth-contract-config.json
-solana-contract-config.json
+solana-program-config.json
 docker-compose/.development.env
 .nyc_output/
 coverage/
